Fix first operation bug
Fixed a bug where the display would not return the operation of two numbers when an operator is pressed next. Whenever a user clicks an operator after a first number and second number is initialized, the operation of those two will become the first number and be displayed along with the newly pressed value.